fix typecast warning

+++ src/chatbot.c
@@ -54,7 +54,7 @@ hashtable_t *ht_create( int size ) {
 int ht_hash( hashtable_t *hashtable, char *key ) {
 
   unsigned long int hashval = 0;
-  int i = 0;
+  unsigned long i = 0;
 
   /* Convert our string to an integer */
   while( hashval < ULONG_MAX && i < strlen( key ) ) {
